[balances-tx-payment-extension] separate aliases from conversion
